+++ CHANGES.md
@@ -27,6 +27,9 @@
 - Body and tail modules not correctly freezing
 - Made `Swish` to not be inplace - seemed to cause problems sometimes
 - Enforced fastprogress version; latest version renamed a parameter
+- Added support to `df2foldfile` for missing `strat_key`
+- Added support to `fold2foldfile` for missing features
+
 
 ## Changes